code refactoring.
